implement test api images history integration test on dedicated file
(cherry picked from commit f6b3b257c828835f4a3bca5b9eb95edcf76e8946)
